pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "pp-nextgen"
version = "0.1.0"
description = "Fine-grained heterogeneous pipeline-parallel inference (nextgen)"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "numpy>=1.24",
  "PyYAML>=6.0",
  "matplotlib>=3.8",
  "grpcio>=1.62.0",
  "protobuf>=4.25.0",
]

[project.optional-dependencies]
dev = ["ruff>=0.4", "grpcio-tools>=1.62.0"]

[tool.setuptools.packages.find]
where = ["."]
include = ["pp_nextgen*"]
